complete generator script
mypy runs without errors on the generated stubs now. The only thing that
is left is to refine the type annotations in the docstrings a bit. The
main task is to get some nice base class for enumerations.
