+++ scratch/mypy/abc.py
@@ -1,13 +1,3 @@
-# pylint: disable=invalid-name, missing-module-docstring, unused-argument, missing-function-docstring, dangerous-default-value, line-too-long
-# pylint: disable=multiple-statements, bad-whitespace, function-redefined, missing-class-docstring, unsubscriptable-object, redefined-builtin
-
-# we can only support python 3.7 but nothing before with this
-from __future__ import annotations
-
-from typing import Any, Callable, ContextManager, Iterator, List, Mapping, MutableSequence, Optional, Sequence, Tuple, Union
-from abc import ABC, ABCMeta, abstractmethod
-
-
 class Application(ABC):
     """
     Interface that has to be implemented to customize clingo.
@@ -607,246 +597,3 @@ class Propagator(ABC):
         decision. If all propagators return 0, then the fallback literal is
         used.
         """
-
-Infimum: Symbol
-Supremum: Symbol
-__version__: str
-
-
-def Function(name: str, arguments: List[Symbol]=[], positive: bool=True) -> Symbol: ...
-def Number(number: int) -> Symbol: ...
-def String(string: str) -> Symbol: ...
-def Tuple_(arguments: List[Symbol]) -> Symbol: ...
-def clingo_main(application: Application, files: List[str]=[]) -> int: ...
-def parse_program(program: str, callback: Callable[[ast.AST], None]) -> None: ...
-def parse_term(string: str, logger: Callable[[MessageCode,str],None]=None, message_limit: int=20) -> Symbol: ...
-
-
-class ApplicationOptions(metaclass=ABCMeta):
-    def add(self, group: str, option: str, description: str, parser: Callable[[str], bool], multi: bool=False, argument: str=None) -> None: ...
-    def add_flag(self, group: str, option: str, description: str, target: Flag) -> None: ...
-
-class Assignment(Sequence[int], metaclass=ABCMeta):
-    def decision(self, level: int) -> int: ...
-    def has_literal(self, literal : int) -> bool: ...
-    def is_false(self, literal: int) -> bool: ...
-    def is_fixed(self, literal: int) -> bool: ...
-    def is_true(self, literal: int) -> bool: ...
-    def level(self, literal: int) -> int: ...
-    def value(self, literal) -> Optional[bool]: ...
-    decision_level: int
-    has_conflict: bool
-    is_total: bool
-    root_level: int
-    trail: Trail
-
-class Backend(ContextManager[Backend], metaclass=ABCMeta):
-    def add_acyc_edge(self, node_u: int, node_v: int, condition: List[int]) -> None: ...
-    def add_assume(self, literals: List[int]) -> None: ...
-    def add_atom(self, symbol : Optional[Symbol]=None) -> int: ...
-    def add_external(self, atom : int, value : TruthValue=TruthValue.False_) -> None: ...
-    def add_heuristic(self, atom: int, type: HeuristicType, bias: int, priority: int, condition: List[int]) -> None: ...
-    def add_minimize(self, priority: int, literals: List[Tuple[int,int]]) -> None: ...
-    def add_project(self, atoms: List[int]) -> None: ...
-    def add_rule(self, head: List[int], body: List[int]=[], choice: bool=False) -> None: ...
-    def add_weight_rule(self, head: List[int], lower: int, body: List[Tuple[int,int]], choice: bool=False) -> None: ...
-
-class Configuration(metaclass=ABCMeta):
-    def description(self, name: str) -> str: ...
-    keys: Optional[List[str]]
-
-class Control:
-    def __init__(self, arguments: List[str]=[], logger: Callable[[MessageCode,str],None]=None, message_limit: int=20): ...
-    def add(self, name: str, parameters: List[str], program: str) -> None: ...
-    def assign_external(self, external: Union[Symbol,int], truth: Optional[bool]) -> None: ...
-    def backend(self) -> Backend: ...
-    def builder(self) -> ProgramBuilder: ...
-    def cleanup(self) -> None: ...
-    def get_const(self, name: str) -> Optional[Symbol]: ...
-    def ground(self, parts: List[Tuple[str,List[Symbol]]], context: Any=None) -> None: ...
-    def interrupt(self) -> None: ...
-    def load(self, path: str) -> None: ...
-    def register_observer(self, observer: Observer, replace: bool=False) -> None: ...
-    def register_propagator(self, propagator: Propagator) -> None: ...
-    def release_external(self, symbol: Union[Symbol,int]) -> None: ...
-    def solve(self, assumptions: List[Union[Tuple[Symbol,bool],int]]=[], on_model: Callable[[Model],Optional[bool]]=None, on_statistics : Callable[[StatisticsMap,StatisticsMap],None]=None, on_finish: Callable[[SolveResult],None]=None, yield_: bool=False, async_: bool=False) -> Union[SolveHandle,SolveResult]: ...
-    configuration: Configuration
-    is_conflicting: bool
-    statistics: dict
-    symbolic_atoms: SymbolicAtoms
-    theory_atoms: TheoryAtomIter
-    set_use_enumeration_assumption: bool
-
-class Flag:
-    def __init__(self, value: bool=False): ...
-    flag: bool
-
-class HeuristicType(metaclass=ABCMeta):
-    Factor: HeuristicType
-    False_: HeuristicType
-    Init: HeuristicType
-    Level: HeuristicType
-    Sign: HeuristicType
-    True_: HeuristicType
-
-class MessageCode(metaclass=ABCMeta):
-    AtomUndefined: MessageCode
-    FileIncluded: MessageCode
-    GlobalVariable: MessageCode
-    OperationUndefined: MessageCode
-    Other: MessageCode
-    RuntimeError: MessageCode
-    VariableUnbounded: MessageCode
-
-class Model(metaclass=ABCMeta):
-    def contains(self, atom: Symbol) -> bool: ...
-    def extend(self, symbols: List[Symbol]) -> None: ...
-    def is_true(self, literal: int) -> bool: ...
-    def symbols(self, atoms: bool=False, terms: bool=False, shown: bool=False, csp: bool=False, theory: bool=False, complement: bool=False) -> List[Symbol]: ...
-    context: SolveControl
-    cost: List[int]
-    number: int
-    optimality_proven: bool
-    thread_id: int
-    type: ModelType
-
-class ModelType(metaclass=ABCMeta):
-    BraveConsequences: ModelType
-    CautiousConsequences: ModelType
-    StableModel: ModelType
-
-class ProgramBuilder(ContextManager[ProgramBuilder], metaclass=ABCMeta):
-    def add(self, statement: ast.AST) -> None: ...
-
-class PropagateControl(metaclass=ABCMeta):
-    def add_clause(self, clause: List[int], tag: bool=False, lock: bool=False) -> bool: ...
-    def add_literal(self) -> int: ...
-    def add_nogood(self, clause: List[int], tag: bool=False, lock: bool=False) -> bool: ...
-    def add_watch(self, literal: int) -> None: ...
-    def has_watch(self, literal: int) -> bool: ...
-    def propagate(self) -> bool: ...
-    def remove_watch(self, literal: int) -> None: ...
-    assignment: Assignment
-    thread_id: int
-
-class PropagateInit(metaclass=ABCMeta):
-    def add_clause(self, clause: List[int]) -> bool: ...
-    def add_literal(self) -> int: ...
-    def add_minimize(self, literal: int, weight: int, priority: int=0) -> None: ...
-    def add_watch(self, literal: int, thread_id: Optional[int]=None) -> None: ...
-    def add_weight_constraint(self, literal: int, literals: List[Tuple[int,int]], bound: int, type: int=0, compare_equal: bool=False) -> bool: ...
-    def propagate(self) -> bool: ...
-    def solver_literal(self, literal: int) -> int: ...
-    assignment: Assignment
-    check_mode: PropagatorCheckMode
-    number_of_threads: int
-    symbolic_atoms: SymbolicAtoms
-    theory_atom: TheoryAtomIter
-
-class PropagatorCheckMode(metaclass=ABCMeta):
-    Both: PropagatorCheckMode
-    Fixpoint: PropagatorCheckMode
-    Off: PropagatorCheckMode
-    Total: PropagatorCheckMode
-
-class Slice(metaclass=ABCMeta):
-    pass
-
-class SolveControl(metaclass=ABCMeta):
-    def add_clause(self, literals: List[Union[Tuple[Symbol,bool],int]]) -> None: ...
-    def add_nogood(self, literals: List[Union[Tuple[Symbol,bool],int]]) -> None: ...
-    symbolic_atoms: SymbolicAtoms
-
-class SolveHandle(ContextManager[SolveHandle], metaclass=ABCMeta):
-    def cancel(self) -> None: ...
-    def core(self) -> List[int]: ...
-    def get(self) -> SolveResult: ...
-    def resume(self) -> None: ...
-    def wait(self, timeout: Optional[float]=None) -> Optional[bool]: ...
-
-class SolveResult(metaclass=ABCMeta):
-    exhausted: bool
-    interruped: bool
-    satisfiable: Optional[bool]
-    unknown: bool
-    unsatisfiable: Optional[bool]
-
-class StatisticsArray(MutableSequence[Union[StatisticsArray,StatisticsMap,float]], metaclass=ABCMeta):
-    def append(self, value: Any) -> None: ...
-    def extend(self, values: Sequence[Any]) -> None: ...
-    def update(self, values: Sequence[Any]) -> None: ...
-
-class StatisticsMap(Mapping[str,Union[StatisticsArray,StatisticsMap,float]], metaclass=ABCMeta):
-    def items(self) -> List[Tuple[str, Union[StatisticsArray,StatisticsMap,float]]]: ...
-    def keys(self) -> List[str]: ...
-    def update(self, values: Mapping[str,Any]) -> None: ...
-    def values(self) -> List[Union[StatisticsArray,StatisticsMap,float]]: ...
-
-class Symbol(metaclass=ABCMeta):
-    def match(self, name: str, arity: int) -> bool: ...
-    arguments: List[Symbol]
-    name: str
-    negative: bool
-    number: int
-    positive: bool
-    string: str
-    type: SymbolType
-
-class SymbolType(metaclass=ABCMeta):
-    Function: SymbolType
-    Infimum: SymbolType
-    Number: SymbolType
-    String: SymbolType
-    Supremum: SymbolType
-
-class SymbolicAtom(metaclass=ABCMeta):
-    def match(self, name: str, arity: int) -> bool: ...
-    is_external: bool
-    is_fact: bool
-    literal: int
-    symbol: Symbol
-
-class SymbolicAtomIter(Iterator[SymbolicAtom], metaclass=ABCMeta):
-    pass
-
-class SymbolicAtoms(Mapping[Union[Symbol,int],SymbolicAtom], metaclass=ABCMeta):
-    def by_signature(self, name: str, arity: int, positive: bool=True) -> Iterator[Symbol]: ...
-    signatures: List[Tuple[str,int,bool]]
-
-class TheoryAtom(metaclass=ABCMeta):
-    elements: List[TheoryElement]
-    guard: Tuple[str, TheoryTerm]
-    literal: int
-    term: TheoryTerm
-
-class TheoryAtomIter(Iterator[TheoryAtom], metaclass=ABCMeta):
-    pass
-
-class TheoryElement(metaclass=ABCMeta):
-    condition: List[TheoryTerm]
-    condition_id: int
-    terms: List[TheoryTerm]
-
-class TheoryTerm(metaclass=ABCMeta):
-    arguments: List[TheoryTerm]
-    name: str
-    number: int
-    type: TheoryTermType
-
-class TheoryTermType(metaclass=ABCMeta):
-    Function: TheoryTermType
-    List: TheoryTermType
-    Number: TheoryTermType
-    Set: TheoryTermType
-    Symbol: TheoryTermType
-    Tuple: TheoryTermType
-
-class Trail(Sequence[int], metaclass=ABCMeta):
-    def begin(self, level: int) -> int: ...
-    def end(self, level: int) -> int: ...
-
-class TruthValue(metaclass=ABCMeta):
-    False_: TruthValue
-    Free: TruthValue
-    Release: TruthValue
-    True_: TruthValue
